add more test coverage
